move json parsing logic to ContractStorageSchema
