+++ src/bctklib/models/ContractTypes.cs
@@ -2,8 +2,6 @@ using System.Collections.Generic;
 
 namespace Neo.BlockchainToolkit.Models
 {
-    public abstract record ContractType();
-
     public enum PrimitiveType : byte
     {
         Boolean,
@@ -17,9 +15,11 @@ namespace Neo.BlockchainToolkit.Models
         Address,
     }
 
+    public abstract record ContractType();
+
     public record PrimitiveContractType(PrimitiveType Type) : ContractType;
 
-    public record StructContractType(IReadOnlyList<(string Name, ContractType Type)> Fields) : ContractType;
+    public record StructContractType(string Name, IReadOnlyList<(string Name, ContractType Type)> Fields) : ContractType;
 
     public record MapContractType(PrimitiveType KeyType, ContractType ValueType) : ContractType;
 }
